Move TYPES to constants file
