+++ frontend/src/metabase-lib/lib/queries/drills/column-filter-drill.js
@@ -1,4 +1,5 @@
-import { isa, TYPE } from "metabase-lib/lib/types/utils/isa";
+import { isa } from "metabase-lib/lib/types/utils/isa";
+import { TYPE } from "metabase-lib/lib/types/constants";
 
 const INVALID_TYPES = [TYPE.Structured];
