+++ frontend/src/metabase-lib/lib/queries/drills/distribution-drill.js
@@ -1,4 +1,5 @@
-import { isa, TYPE } from "metabase-lib/lib/types/utils/isa";
+import { isa } from "metabase-lib/lib/types/utils/isa";
+import { TYPE } from "metabase-lib/lib/types/constants";
 
 const DENYLIST_TYPES = [
   TYPE.PK,
