+++ frontend/src/metabase-lib/lib/queries/drills/sort-drill.js
@@ -1,4 +1,5 @@
-import { isa, TYPE } from "metabase-lib/lib/types/utils/isa";
+import { isa } from "metabase-lib/lib/types/utils/isa";
+import { TYPE } from "metabase-lib/lib/types/constants";
 import Dimension from "metabase-lib/lib/Dimension";
 
 const INVALID_TYPES = [TYPE.Structured];
