+++ frontend/src/metabase-lib/lib/queries/structured/Aggregation.ts
@@ -5,7 +5,7 @@ import { Aggregation as AggregationObject } from "metabase-types/types/Query";
 import { AggregationOperator } from "metabase-types/types/Metadata";
 import { MetricId } from "metabase-types/types/Metric";
 import { FieldId } from "metabase-types/types/Field";
-import { TYPE } from "metabase-lib/lib/types/utils/isa";
+import { TYPE } from "metabase-lib/lib/types/constants";
 import * as AGGREGATION from "metabase-lib/lib/queries/utils/aggregation";
 import Filter from "metabase-lib/lib/queries/structured/Filter";
 import Metric from "metabase-lib/lib/metadata/Metric";
