+++ frontend/src/metabase-lib/lib/queries/utils/drilldown.js
@@ -1,6 +1,7 @@
 import _ from "underscore";
 import { isLatitude, isLongitude, isDate } from "metabase/lib/schema_metadata";
-import { isa, TYPE } from "metabase-lib/lib/types/utils/isa";
+import { isa } from "metabase-lib/lib/types/utils/isa";
+import { TYPE } from "metabase-lib/lib/types/constants";
 
 import { isExpressionField } from "metabase-lib/lib/queries/utils/field-ref";
 import { FieldDimension } from "metabase-lib/lib/Dimension";
