+++ frontend/src/metabase-lib/lib/types/constants.js
@@ -1,0 +1,3 @@
+import { TYPE as cljs_TYPE } from "cljs/metabase.types";
+
+export const TYPE = cljs_TYPE;
