+++ frontend/src/metabase/admin/settings/components/widgets/FormattingWidget.jsx
@@ -3,7 +3,7 @@ import React from "react";
 import { t } from "ttag";
 
 import ColumnSettings from "metabase/visualizations/components/ColumnSettings";
-import { TYPE } from "metabase-lib/lib/types/utils/isa";
+import { TYPE } from "metabase-lib/lib/types/constants";
 
 const SETTING_TYPES = [
   {
