+++ frontend/src/metabase/entities/fields.js
@@ -23,7 +23,7 @@ import {
   field_semantic_types,
   has_field_values_options,
 } from "metabase/lib/core";
-import { TYPE } from "metabase-lib/lib/types/utils/isa";
+import { TYPE } from "metabase-lib/lib/types/constants";
 import {
   getFieldValues,
   getRemappings,
