+++ frontend/src/metabase/lib/core.js
@@ -1,5 +1,5 @@
 import { t } from "ttag";
-import { TYPE } from "metabase-lib/lib/types/utils/isa";
+import { TYPE } from "metabase-lib/lib/types/constants";
 
 export const field_semantic_types = [
   /* Overall Row */
