+++ frontend/src/metabase/visualizations/components/ObjectDetail/ObjectDetailsTable.tsx
@@ -9,7 +9,8 @@ import EmptyState from "metabase/components/EmptyState";
 
 import { isID } from "metabase/lib/schema_metadata";
 import { formatValue, formatColumn } from "metabase/lib/formatting";
-import { TYPE, isa } from "metabase-lib/lib/types/utils/isa";
+import { isa } from "metabase-lib/lib/types/utils/isa";
+import { TYPE } from "metabase-lib/lib/types/constants";
 
 import { OnVisualizationClickType } from "./types";
 import {
