+++ frontend/test/metabase/reference/utils.unit.spec.js
@@ -1,7 +1,7 @@
 import { databaseToForeignKeys, getQuestion } from "metabase/reference/utils";
 
 import { separateTablesBySchema } from "metabase/reference/databases/TableList";
-import { TYPE } from "metabase-lib/lib/types/utils/isa";
+import { TYPE } from "metabase-lib/lib/types/constants";
 
 describe("Reference utils.js", () => {
   describe("databaseToForeignKeys()", () => {
