+++ frontend/test/metabase/visualizations/lib/table.unit.spec.js
@@ -3,7 +3,7 @@ import {
   getTableClickedObjectRowData,
   isColumnRightAligned,
 } from "metabase/visualizations/lib/table";
-import { TYPE } from "metabase-lib/lib/types/utils/isa";
+import { TYPE } from "metabase-lib/lib/types/constants";
 
 const RAW_COLUMN = {
   source: "fields",
